pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Professional <pyostr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.5"
aiogram = "^3.22.0"
python-dotenv = "^1.1.1"
redis = "^6.4.0"
boto3 = "^1.40.50"
pillow = "^11.3.0"
httplib2 = "^0.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
